pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["inbasu <just.me.juncy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
wfastcgi = "^3.0.0"
Django = "^5.0.6"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
logging = "^0.4.9.6"
numpy = "^1.26.4"
eventlet = "^0.37.0"
requests = "^2.32.3"
celery = "^5.4.0"
types-requests = "^2.32.0.20240914"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
python-dotenv = "^1.0.1"
pytest = "^8.2.1"
ruff = "^0.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "fail", "succ",
    "zebra_it",
]
